AI-Dementia-Diagnosis/requirements.txt
tensorflow-cpu==2.15.0
nibabel==5.2.0
scikit-image==0.22.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
pyyaml==6.0.1
tqdm==4.66.1
openpyxl==3.1.2